Conflicts:
	.gitignore
	gwt-unitCache/gwt-unitCache-b43fe07841b100c5f8cde7d5e855fc294f6a56f4-0000014688726174
	gwt-unitCache/gwt-unitCache-b43fe07841b100c5f8cde7d5e855fc294f6a56f4-000001468CC44072
	gwt-unitCache/gwt-unitCache-b43fe07841b100c5f8cde7d5e855fc294f6a56f4-000001468DC127DA
	gwt-unitCache/gwt-unitCache-b43fe07841b100c5f8cde7d5e855fc294f6a56f4-000001468DD8B308
	gwt-unitCache/gwt-unitCache-b43fe07841b100c5f8cde7d5e855fc294f6a56f4-000001468DE23B09

+++ src/com/ubc/cpsc310/vancouverparking/client/MeterCell.java
@@ -6,6 +6,7 @@ import com.google.gwt.safecss.shared.SafeStyles;
 import com.google.gwt.safehtml.client.SafeHtmlTemplates;
 import com.google.gwt.safehtml.shared.SafeHtml;
 import com.google.gwt.safehtml.shared.SafeHtmlBuilder;
+import com.ubc.cpsc310.vancouverparking.server.Meter;
 
 public class MeterCell extends AbstractCell<MeterInfo>{
 	
@@ -28,7 +29,7 @@ public class MeterCell extends AbstractCell<MeterInfo>{
       		+ "<span class='meter_number'>{0}</span><br>"
     		+ "<span class='meter_data'>{1} | {2}</span>"
       		+ "</div>")
-      SafeHtml cell(int meterNum, double rate, String paymentType, String iconStyle);
+      SafeHtml cell(long meterNum, double rate, String paymentType, String iconStyle);
     }
     
     
@@ -53,7 +54,7 @@ public class MeterCell extends AbstractCell<MeterInfo>{
 	      }
 	      
 	      // Use the template to create the Cell's html.
-	      int meterNum = value.getNumber();
+	      long meterNum = value.getNumber();
 	      double rate = value.getRate();
 	      String paymentType;
 	      if (value.isCreditCard()) {
